Fix uploading resources

+++ upload-installation-manager.sh
@@ -115,27 +115,29 @@ uploadCodenvyMultiServerInstallProperties() {
 }
 
 doUpload() {
-    DESTINATION=update-server-repository/${ARTIFACT}/${VERSION}
+    if [[ -f ${SOURCE} ]]; then
+        DESTINATION=update-server-repository/${ARTIFACT}/${VERSION}
 
-    MD5=`md5sum ${SOURCE} | cut -d ' ' -f 1`
-    SIZE=`du -b ${SOURCE} | cut -f1`
-    BUILD_TIME=`stat -c %y ${SOURCE}`
-    BUILD_TIME=${BUILD_TIME:0:19}
+        MD5=`md5sum ${SOURCE} | cut -d ' ' -f 1`
+        SIZE=`du -b ${SOURCE} | cut -f1`
+        BUILD_TIME=`stat -c %y ${SOURCE}`
+        BUILD_TIME=${BUILD_TIME:0:19}
 
-    echo "file=${FILENAME}" > .properties
-    echo "artifact=${ARTIFACT}" >> .properties
-    echo "version=${VERSION}" >> .properties
-    echo "description=${DESCRIPTION}" >> .properties
-    echo "authentication-required=false" >> .properties
-    echo "build-time="${BUILD_TIME} >> .properties
-    echo "md5=${MD5}" >> .properties
-    echo "size=${SIZE}" >> .properties
+        echo "file=${FILENAME}" > .properties
+        echo "artifact=${ARTIFACT}" >> .properties
+        echo "version=${VERSION}" >> .properties
+        echo "description=${DESCRIPTION}" >> .properties
+        echo "authentication-required=false" >> .properties
+        echo "build-time="${BUILD_TIME} >> .properties
+        echo "md5=${MD5}" >> .properties
+        echo "size=${SIZE}" >> .properties
 
-    ssh -i ${SSH_KEY_NAME} ${SSH_AS_USER_NAME}@${AS_IP} "mkdir -p /home/${SSH_AS_USER_NAME}/${DESTINATION}"
-    scp -o StrictHostKeyChecking=no -i ${SSH_KEY_NAME} ${SOURCE} ${SSH_AS_USER_NAME}@${AS_IP}:${DESTINATION}/${FILENAME}
-    scp -o StrictHostKeyChecking=no -i ${SSH_KEY_NAME} .properties ${SSH_AS_USER_NAME}@${AS_IP}:${DESTINATION}/.properties
+        ssh -i ${SSH_KEY_NAME} ${SSH_AS_USER_NAME}@${AS_IP} "mkdir -p /home/${SSH_AS_USER_NAME}/${DESTINATION}"
+        scp -o StrictHostKeyChecking=no -i ${SSH_KEY_NAME} ${SOURCE} ${SSH_AS_USER_NAME}@${AS_IP}:${DESTINATION}/${FILENAME}
+        scp -o StrictHostKeyChecking=no -i ${SSH_KEY_NAME} .properties ${SSH_AS_USER_NAME}@${AS_IP}:${DESTINATION}/.properties
 
-    rm .properties
+        rm .properties
+    fi
 }
 
 uploadInstallationManagerCli
